vexriscv: downgrade litex to working version
The latest Litex does not boot on the VexRiscV, if it is build with
Yocto. Downgrade to the same version as the Rocket CPU, because it
works.

This seems to be related to the Risc-V toolchain, because using the
prebuilt sifive toolchain [0] results in a working bios. The bitstreams
differ only marginally between the Yocto build and directly using LiteX.

[0] riscv64-unknown-elf-gcc-8.1.0-2019.01.0-x86_64-linux-ubuntu14.tar.gz