ecpix5-vexriscv: Add core-image-minimal-ecpix5-vexriscv.cpio to SD card
Mounting the SD card in Linux does not work, yet. Therefore, we need to
mount the rootfs as a cpio, which is loaded by the bios (which can
access the SD card).

However, the dependencies to add the cpio to the SD card don't work
either. Therefore, require a manual step to copy the cpio to the sd
card, but prepare everything else to be in place.