fpga-software: synthesis: litex: autogenerate devicetree
Adding IP cores to litex changes the memory addresses of the other
cores. Therefore a static devicetree is always at danger to be
out-dated. Use the litex_json2dts_linux script to generate the devictree
from the generated memory map of litex.

Changes to this devicetree are probably best handled via overlays.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>